refactoring common file name
