+++ src/cfnlint/rules/common.py
