+++ src/cfnlint/rules/mappings/Name.py
@@ -3,7 +3,7 @@ Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: MIT-0
 """
 from cfnlint.rules import CloudFormationLintRule
-from cfnlint.rules.limits import name
+from cfnlint.rules.common import name
 
 
 class Name(CloudFormationLintRule):
